LL with Auto
ColorLLB&R work very well they are just to test that our pipelines are working and that we are identifying what we should be. Trajectories are all the new ones for my AuBlueLL file. This is my two specimen auto one with the insurance plan for if we are not able to grab a specimen from the human player the LL will detect it and move on. It is a very sketchy/questionable code.
LL = Limelight
Red = Pipeline 0
Blue = Pipeline 2