Auto Changes
Organize Auto files on the DS using the grouping, fixed some of my spelling, worked on the limelight in auto. It is still not cooroporating and I don't know why so I am putting all of the logic in one file and that did not fix it. I got no clue at this point what is wrong with it.